TO-DO-LIST-APPLICATION
A task application developed with python that includes tools for adding, deleting, listing, editing tasks, prints the added tasks to a txt file and you can perform operations according to the selection. In addition, the application checks the data entered with try, except methods and outputs in a numbered way.